Split out tests and try to use the wasm_bindgen_test attribute
